Move pages into their respective directories
