+++ docs/src/pages/docs/tables/index.jsx
@@ -1,8 +1,10 @@
 import { UsaceBox } from "@usace/groundwork";
 import {TSTable} from "@usace-watermanagement/groundwork-water";
+import DocsPage from "../_docs-wrapper";
 
 function Tables() {
   return (
+    <DocsPage>
     <div className="mt-6">
       <UsaceBox title="Welcome">
         <TSTable
@@ -14,6 +16,7 @@ function Tables() {
         />
       </UsaceBox>
     </div>
+    </DocsPage>
   );
 }
